pyproject.toml
[project]
name = "linkedin-post-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.6",
    "langchain-groq>=1.1.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
]
